[Fluid] Moved LuaJIT parser files to their own folder
