+++ src/fluid/luajit-2.1/src/lib/lj_lib.cpp
@@ -22,7 +22,7 @@
 #include "lj_vm.h"
 #include "lj_strscan.h"
 #include "lj_strfmt.h"
-#include "lj_lex.h"
+#include "../parser/lj_lex.h"
 #include "lj_bcdump.h"
 #include "lj_lib.h"
