+++ src/fluid/luajit-2.1/src/lj_asm_arm64.h
@@ -697,7 +697,7 @@ static void asm_strto(ASMState* as, IRIns* ir)
       emit_lso(as, A64I_LDRd, (dest & 31), RID_SP, 0);
    asm_guardcnb(as, A64I_CBZ, RID_RET);
    args[0] = ir->op1; //  GCstr *str
-   args[1] = ASMREF_TMP1; //  TValue *n 
+   args[1] = ASMREF_TMP1; //  TValue *n
    asm_gencall(as, ci, args);
    tmp = ra_releasetmp(as, ASMREF_TMP1);
    emit_opk(as, A64I_ADDx, tmp, RID_SP, ofs, RSET_GPR);
@@ -1331,8 +1331,8 @@ static void asm_cnew(ASMState* as, IRIns* ir)
    else if (ir->op2 != REF_NIL) {  // Create VLA/VLS/aligned cdata.
       ci = &lj_ir_callinfo[IRCALL_lj_cdata_newv];
       args[0] = ASMREF_L;     //  lua_State *L
-      args[1] = ir->op1;      //  CTypeID id  
-      args[2] = ir->op2;      //  CTSize sz   
+      args[1] = ir->op1;      //  CTypeID id
+      args[2] = ir->op2;      //  CTSize sz
       args[3] = ASMREF_TMP1;  //  CTSize align
       asm_gencall(as, ci, args);
       emit_loadi(as, ra_releasetmp(as, ASMREF_TMP1), (int32_t)ctype_align(info));
@@ -1348,7 +1348,7 @@ static void asm_cnew(ASMState* as, IRIns* ir)
       if (id < 65536) emit_d(as, A64I_MOVZw | A64F_U16(id), RID_X1);
    }
    args[0] = ASMREF_L;     //  lua_State *L
-   args[1] = ASMREF_TMP1;  //  MSize size  
+   args[1] = ASMREF_TMP1;  //  MSize size
    asm_gencall(as, ci, args);
    ra_allockreg(as, (int32_t)(sz + sizeof(GCcdata)),
       ra_releasetmp(as, ASMREF_TMP1));
@@ -1385,7 +1385,7 @@ static void asm_obar(ASMState* as, IRIns* ir)
    ra_evictset(as, RSET_SCRATCH);
    l_end = emit_label(as);
    args[0] = ASMREF_TMP1;  //  global_State *g
-   args[1] = ir->op1;      //  TValue *tv     
+   args[1] = ir->op1;      //  TValue *tv
    asm_gencall(as, ci, args);
    emit_dm(as, A64I_MOVx, ra_releasetmp(as, ASMREF_TMP1), RID_GL);
    obj = IR(ir->op1)->r;
@@ -1939,7 +1939,7 @@ static void asm_gc_check(ASMState* as)
    asm_guardcnb(as, A64I_CBNZ, RID_RET); //  Assumes asm_snap_prep() is done.
    *--as->mcp = ARM64_NOPATCH_GC_CHECK;
    args[0] = ASMREF_TMP1;  //  global_State *g
-   args[1] = ASMREF_TMP2;  //  MSize steps    
+   args[1] = ASMREF_TMP2;  //  MSize steps
    asm_gencall(as, ci, args);
    emit_dm(as, A64I_MOVx, ra_releasetmp(as, ASMREF_TMP1), RID_GL);
    tmp2 = ra_releasetmp(as, ASMREF_TMP2);
