+++ src/fluid/luajit-2.1/src/lj_asm_ppc.h
@@ -601,7 +601,7 @@ static void asm_strto(ASMState* as, IRIns* ir)
    asm_guardcc(as, CC_EQ);
    emit_ai(as, PPCI_CMPWI, RID_RET, 0);  //  Test return status.
    args[0] = ir->op1;      //  GCstr *str
-   args[1] = ASMREF_TMP1;  //  TValue *n 
+   args[1] = ASMREF_TMP1;  //  TValue *n
    asm_gencall(as, ci, args);
    // Store the result to the spill slot or temp slots.
    emit_tai(as, PPCI_ADDI, ra_releasetmp(as, ASMREF_TMP1), RID_SP, ofs);
@@ -1276,8 +1276,8 @@ static void asm_cnew(ASMState* as, IRIns* ir)
    else if (ir->op2 != REF_NIL) {  // Create VLA/VLS/aligned cdata.
       ci = &lj_ir_callinfo[IRCALL_lj_cdata_newv];
       args[0] = ASMREF_L;     //  lua_State *L
-      args[1] = ir->op1;      //  CTypeID id  
-      args[2] = ir->op2;      //  CTSize sz   
+      args[1] = ir->op1;      //  CTypeID id
+      args[2] = ir->op2;      //  CTSize sz
       args[3] = ASMREF_TMP1;  //  CTSize align
       asm_gencall(as, ci, args);
       emit_loadi(as, ra_releasetmp(as, ASMREF_TMP1), (int32_t)ctype_align(info));
@@ -1290,7 +1290,7 @@ static void asm_cnew(ASMState* as, IRIns* ir)
    emit_ti(as, PPCI_LI, RID_RET + 1, ~LJ_TCDATA);
    emit_ti(as, PPCI_LI, RID_TMP, id);  //  Lower 16 bit used. Sign-ext ok.
    args[0] = ASMREF_L;     //  lua_State *L
-   args[1] = ASMREF_TMP1;  //  MSize size  
+   args[1] = ASMREF_TMP1;  //  MSize size
    asm_gencall(as, ci, args);
    ra_allockreg(as, (int32_t)(sz + sizeof(GCcdata)),
       ra_releasetmp(as, ASMREF_TMP1));
@@ -1327,7 +1327,7 @@ static void asm_obar(ASMState* as, IRIns* ir)
    ra_evictset(as, RSET_SCRATCH);
    l_end = emit_label(as);
    args[0] = ASMREF_TMP1;  //  global_State *g
-   args[1] = ir->op1;      //  TValue *tv     
+   args[1] = ir->op1;      //  TValue *tv
    asm_gencall(as, ci, args);
    emit_tai(as, PPCI_ADDI, ra_releasetmp(as, ASMREF_TMP1), RID_JGL, -32768);
    obj = IR(ir->op1)->r;
@@ -2211,7 +2211,7 @@ static void asm_gc_check(ASMState* as)
    *--as->mcp = PPC_NOPATCH_GC_CHECK;
    emit_ai(as, PPCI_CMPWI, RID_RET, 0);
    args[0] = ASMREF_TMP1;  //  global_State *g
-   args[1] = ASMREF_TMP2;  //  MSize steps    
+   args[1] = ASMREF_TMP2;  //  MSize steps
    asm_gencall(as, ci, args);
    emit_tai(as, PPCI_ADDI, ra_releasetmp(as, ASMREF_TMP1), RID_JGL, -32768);
    tmp = ra_releasetmp(as, ASMREF_TMP2);
