+++ src/fluid/luajit-2.1/src/lj_ffrecord.h
@@ -3,8 +3,7 @@
 ** Copyright (C) 2005-2022 Mike Pall. See Copyright Notice in luajit.h
 */
 
-#ifndef _LJ_FFRECORD_H
-#define _LJ_FFRECORD_H
+#pragma once
 
 #include "lj_obj.h"
 #include "lj_jit.h"
@@ -19,6 +18,4 @@ typedef struct RecordFFData {
 
 LJ_FUNC int32_t lj_ffrecord_select_mode(jit_State* J, TRef tr, TValue* tv);
 LJ_FUNC void lj_ffrecord_func(jit_State* J);
-#endif
-
 #endif
