+++ src/fluid/luajit-2.1/src/runtime/lj_state.cpp
@@ -26,7 +26,7 @@
 #include "lj_dispatch.h"
 #include "lj_vm.h"
 #include "lj_prng.h"
-#include "lj_lex.h"
+#include "../parser/lj_lex.h"
 #include "lj_alloc.h"
 #include "luajit.h"
